run flake8 on all test modularize model creation
