+++ test/test_inconsistency_checks.py
@@ -7,7 +7,6 @@ Created on Mon Feb 28 15:47:32 2022.
 """
 
 import numpy as np
-import pandas as pd
 import pytest
 
 from linopy import Model
