+++ test/test_io.py
@@ -12,7 +12,7 @@ import xarray as xr
 from xarray.testing import assert_equal
 
 from linopy import Model, available_solvers, read_netcdf
-from linopy.io import float_to_str, int_to_str
+from linopy.io import int_to_str
 
 
 @pytest.fixture
@@ -37,7 +37,7 @@ def test_str_arrays(m):
 def test_str_arrays_chunked():
     m = Model(chunk="auto")
 
-    x = m.add_variables(4, pd.Series([8, 10]))
+    m.add_variables(4, pd.Series([8, 10]))
     y = m.add_variables(0, pd.DataFrame([[1, 2], [3, 4], [5, 6]]).T)
 
     da = int_to_str(y.compute().values)
@@ -47,9 +47,9 @@ def test_str_arrays_chunked():
 def test_str_arrays_with_nans():
     m = Model()
 
-    x = m.add_variables(4, pd.Series([8, 10]), name="x")
+    m.add_variables(4, pd.Series([8, 10]), name="x")
     # now expand the second dimension, expended values of x will be nan
-    y = m.add_variables(0, pd.DataFrame([[1, 2], [3, 4], [5, 6]]), name="y")
+    m.add_variables(0, pd.DataFrame([[1, 2], [3, 4], [5, 6]]), name="y")
     assert m["x"].values[-1] == -1
 
     da = int_to_str(m["x"].values)
@@ -104,7 +104,7 @@ def test_to_gurobipy(m):
 
 
 @pytest.mark.skipif("highs" not in available_solvers, reason="Highspy not installed")
-def test_to_gurobipy(m):
+def test_to_highspy(m):
     m.to_highspy()
